Abstracts away image props functions, adds in desktop for mockups
